+++ coverage-tests/src/test/java/coverage/TestSetup.java
@@ -35,7 +35,6 @@ public class TestSetup extends GlobalSetup {
         eyes.setParentBranchName("master");
         eyes.setBatch(batch);
         eyes.setSaveNewTests(false);
-        eyes.setForceFullPageScreenshot(false);
         String showLogs = System.getenv("APPLITOOLS_SHOW_LOGS");
         String verbose = System.getenv("APPLITOOLS_SHOW_LOGS_VERBOSE");
         if (showLogs != null && showLogs.equals("true")) {