+++ eyes.sdk.core/src/main/java/com/applitools/eyes/TestResultsStatus.java
@@ -4,5 +4,6 @@ public enum TestResultsStatus {
     Passed,
     Unresolved,
     Failed,
-    NotOpened
+    NotOpened,
+    Disabled
 }