+++ eyes.selenium.java/src/test/java/com/applitools/eyes/selenium/TestAccessibility.java
@@ -55,7 +55,7 @@ public class TestAccessibility extends ReportingTestSuite {
         } finally {
             driver.quit();
             eyes.abortAsync();
-            TestResultsSummary allTestResults = runner.getAllTestResults(false);
+            TestResultsSummary allTestResults = runner.getAllTestResults();
 
             Assert.assertEquals(2, allTestResults.getAllResults().length);
             TestResults resultSanity = allTestResults.getAllResults()[0].getTestResults();