+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [3.190.4]
 ### Updated
 - Scrollable offset calculation with updated Helper Library. [Trello 1673](https://trello.com/c/CYbkzXia)
+### Fixed
+- Fixed some bugs in the dom analyzer. [Trello 2414](https://trello.com/c/45tSPCPv)
 
 ## [3.190.3]
 ### Fixed