+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [vNext]
+## [3.190.5]
 ### Updated
 - Updated logs in visual grid tests to have a test id.